pyproject.toml
[tool.poetry]
name = "polynomial"
version = "0.1.0"
description = ""
authors = ["Olivier Corradi <olivier.corradi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
Django = "^4.1.4"
requests = "^2.28.1"
django-jsonform = "^2.18.0"
environs = {extras = ["django"], version = "^9.5.0"}
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
celery = "^5.2.7"
redis = "^4.4.0"
pandas = "^1.5.2"
requests-oauthlib = "^1.3.1"
django-allauth = "^0.52.0"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
django-compressor = "^4.3.1"
django-widget-tweaks = "^1.4.12"
django-extensions = "^3.2.3"
matplotlib = "^3.7.1"
vl-convert-python = "^0.11.2"
newrelic = "^8.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
black = "^22.12.0"
isort = "^5.11.2"
django-stubs = "^1.13.1"
mypy = "^0.991"
types-requests = "^2.28.11"
types-psycopg2 = "^2.9.21"
types-oauthlib = "^3.2.0"
pandas-stubs = "^1.5.2"
honcho = "^1.1.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipython = "^8.14.0"
autoflake = "^2.2.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
